% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bankr.cbi-package.R
\docType{package}
\name{bankr.cbi-package}
\alias{bankr.cbi}
\alias{bankr.cbi-package}
\title{bankr.cbi: Read and Process CBI Bank Statement Files}
\description{
Provides functions to read, validate, and tidy CBI (Corporate Banking Interbancario) format bank statement files commonly used in Italian banking.
}
\author{
\strong{Maintainer}: Andrea Ranzato \email{andrea.ranzato01@gmail.com}

}
\keyword{internal}
