% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ply.R
\name{chr_ply}
\alias{chr_ply}
\alias{lgl_ply}
\title{Type-stable apply}
\usage{
chr_ply(x, fun, ..., length = 1L, use_names = FALSE)

lgl_ply(x, fun, ..., length = 1L, use_names = FALSE)
}
\arguments{
\item{x, fun, ..., use_names}{Forwarded to \code{\link[base:lapply]{base::vapply()}}}

\item{length}{Result length}
}
\description{
Wrappers for \code{\link[base:lapply]{base::vapply()}} for common types.
}
\keyword{internal}
